+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development v1
 
+### Fixed
+
+- AVPlayer: wrong tracking of startup_time when no source is attached to player
+
 ## 2.0.0
 
 - Support for BitmovinPlayer v3